Extract tapir-cats-effect module from tapir-cats.
