+++ server/armeria-server/cats/src/main/scala/sttp/tapir/server/armeria/cats/CatsMonadAsyncError.scala
@@ -3,7 +3,7 @@ package sttp.tapir.server.armeria.cats
 import cats.effect.Async
 import cats.syntax.functor._
 import sttp.monad.{Canceler, MonadAsyncError}
-import sttp.tapir.integ.cats.CatsMonadError
+import sttp.tapir.integ.cats.effect.CatsMonadError
 
 // Forked from sttp.client3.impl.cats.CatsMonadAsyncError
 private class CatsMonadAsyncError[F[_]](implicit F: Async[F]) extends CatsMonadError[F] with MonadAsyncError[F] {
