+++ server/armeria-server/cats/src/test/scala/sttp/tapir/server/armeria/cats/ArmeriaCatsServerTest.scala
@@ -2,7 +2,7 @@ package sttp.tapir.server.armeria.cats
 
 import cats.effect.{IO, Resource}
 import sttp.capabilities.fs2.Fs2Streams
-import sttp.tapir.integ.cats.CatsMonadError
+import sttp.tapir.integ.cats.effect.CatsMonadError
 import sttp.tapir.server.tests._
 import sttp.tapir.tests.{Test, TestSuite}
